+++ src/wizards/newspack/types/index.d.ts
@@ -96,6 +96,7 @@ declare global {
 				};
 			};
 			'advanced-settings': WizardTab;
+			collections: WizardTab;
 		};
 		newspack_aux_data: {
 			is_debug_mode: boolean;